backend/chatsphere_agent/requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
python-dotenv
pinecone
psycopg2-binary
# Ingestion specific
pymupdf
beautifulsoup4
requests
pytest
